pyproject.toml
[project]
name = "profile-email"
version = "0.1.0"
description = ""
authors = [
    {name = "trighele",email = "trighele@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "resend (>=2.10.0,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "pydantic[email] (>=2.11.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"